+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+- Added `openrpc.json` output: enable it with an `openrpc` feature and `openrpc_outdir` attribute, like this `#[rpc(openrpc_outdir = "./")]`
+- Breaking: you now need to specify that you want typescript bindings as they are not enabled by default `#[rpc(ts_outdir = "typescript/generated")]` instead of just `#[rpc]`
+
 ## 0.4.4
 
 - add `RpcSession::server()` method